Forgot password page
When the user has forgotten his password, he can click on "forgotten password" and will be redirected to the forgot password page.

On it, he will be asked for his username via a form. Once entered, he will receive an email with a link to create a new password which will take him to the reset password page.